requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
kiteconnect>=5.0.0
pandas>=2.1.0
numpy>=1.26.0
SQLAlchemy>=2.0.20
python-dotenv>=1.0.0
httpx>=0.25.0
websockets>=11.0
loguru>=0.7.2
ta>=0.11.0
